+++ Project1/Wall.h
@@ -26,5 +26,5 @@ public:
 
     ~CWall();
 
-    virtual void Draw(Gdiplus::Graphics* graphics, int scroll) override;
+    virtual void Draw(Gdiplus::Graphics* graphics, float scroll) override;
 };
